requirement.txt
bcrypt==3.2.2
blinker==1.5
cffi==1.15.1
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.2.1
Flask==2.2.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.2.0
pycparser==2.21
PyJWT==2.4.0
SQLAlchemy==1.4.39
Werkzeug==2.2.1
WTForms==3.0.1
